Psychology and Counselling
Research of July 2023 l The Role of The School Counseling Process in Sri Lankan Schools Should be Developed 
I conducted an in-depth research project focusing on 'The Role of The School Counseling Process in Sri Lankan Schools.' Through rigorous analysis and dedication to the study, I explored the current state of school counseling in Sri Lanka and identified areas where development is needed. This research highlighted the importance of enhancing the school counseling process to better support students' well-being and academic success. I'm proud to note that my commitment and efforts were recognised with an A+ grade, reflecting the quality and significance of this work.